int instead of double

+++ src/Main.java
@@ -7,13 +7,13 @@ public class Main {
         Input task = new Input();
         task.read();
         boolean res = task.getRomatrue();
-        double result = Calculator.calculate(task.getVar1(), task.getVar2(), task.getOper());
+        int result = Calculator.calculate(task.getVar1(), task.getVar2(), task.getOper());
 
             if (res) {
-                System.out.println("Output:" + "\n" + Backtoroman.intToRoman((int) result));
+                System.out.println("Output:" + "\n" + Backtoroman.intToRoman(result));
             }
             else {
-                System.out.println("Output:" + "\n" + (int) result);
+                System.out.println("Output:" + "\n" + result);
             }
             Start();
